feat: Move custom icons to a separate file
