+++ lib/fluent_ui.dart
@@ -82,7 +82,7 @@ export 'src/controls/surfaces/info_bar.dart';
 export 'src/controls/surfaces/list_tile.dart';
 export 'src/controls/surfaces/progress_indicators.dart';
 export 'src/controls/flyouts/tooltip.dart';
-export 'src/controls/utils/chevron_down.dart';
+export 'src/controls/utils/custom_icons.dart';
 export 'src/controls/utils/divider.dart';
 export 'src/controls/utils/hover_button.dart';
 export 'src/controls/utils/info_badge.dart';
