+++ lib/src/controls/inputs/rating.dart
@@ -308,81 +308,3 @@ enum _SliderAdjustmentType {
   up,
   down,
 }
-
-class RatingIcon extends StatelessWidget {
-  const RatingIcon({
-    super.key,
-    required this.rating,
-    this.ratedColor,
-    this.unratedColor,
-    this.icon = kRatingBarIcon,
-    this.size,
-  }) : assert(rating >= 0.0 && rating <= 1.0);
-
-  /// The rating of the icon. Must be more or equal to 0 and less or equal than 1.0
-  final double rating;
-
-  /// The icon.
-  final IconData icon;
-
-  /// The color used by the rated part. If `null`, uses [FluentThemeData.accentColor]
-  final Color? ratedColor;
-
-  /// The color used by the unrated part. If `null`, uses [FluentThemeData.disabledColor]
-  final Color? unratedColor;
-
-  /// The size of the icon
-  final double? size;
-
-  @override
-  Widget build(BuildContext context) {
-    assert(debugCheckHasFluentTheme(context));
-    final style = FluentTheme.of(context);
-    final icon = this.icon;
-    final size = this.size;
-    final unratedColor =
-        this.unratedColor ?? style.resources.controlFillColorSecondary;
-    final ratedColor =
-        this.ratedColor ?? style.accentColor.defaultBrushFor(style.brightness);
-    if (rating == 1.0) {
-      return Icon(
-        icon,
-        color: ratedColor,
-        size: size,
-      );
-    } else if (rating == 0.0) {
-      return Icon(icon, color: unratedColor, size: size);
-    }
-    return Stack(
-      children: [
-        Icon(icon, color: unratedColor, size: size),
-        ClipRect(
-          clipper: _StarClipper(rating),
-          child: Icon(
-            icon,
-            // IconData(
-            //   fontFamily: 'Segoe MDL2 Assets',
-            // ),
-            color: ratedColor,
-            size: size,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _StarClipper extends CustomClipper<Rect> {
-  final double value;
-
-  _StarClipper(this.value);
-
-  @override
-  Rect getClip(Size size) {
-    final rect = Rect.fromLTWH(0, 0, size.width * value, size.height);
-    return rect;
-  }
-
-  @override
-  bool shouldReclip(_StarClipper oldClipper) => oldClipper.value != value;
-}
